+++ CHANGELOG.md
@@ -1,17 +1,25 @@
 <!--
  * @Author: your name
  * @Date: 2020-03-14 17:21:08
- * @LastEditTime: 2020-05-31 00:15:51
+ * @LastEditTime: 2020-06-06 19:17:54
  * @LastEditors: Please set LastEditors
  * @Description: In User Settings Edit
  * @FilePath: \DevUI-Language-Support\CHANGELOG.md
  -->
+# 2020/6/6 V4.2.1-V4.2.2 
+BUGFix: 
+- 添加了由于将资源树迁移至建筑师类导致的一系列bug
+- 更新wchc脚本
 
+# 2020/6/5 V4.2.0
+	- 添加了对指令的支持。
+	- 将host进行拆分解耦。
+	- 由于种种原因 本版本未使用打包功能。
 # 2020/5/31 V4.1.1
-	bugfix hover将不在于angular无关的项目中触发。
+	- bugfix hover将不在于angular无关的项目中触发。
 # 2020/5/18 V4.1.0
-	增加了Igniter类
-	现在当你处于非angular项目或没有安装devui的项目中，DevUIHelper将不会打扰您
+	- 增加了Igniter类
+	- 现在当你处于非angular项目或没有安装devui的项目中，DevUIHelper将不会打扰您
 # 2020/5/18 V4.0.6
 
 对ast进行了小幅的重写，现在它拥有三个span : namespan tagspan 以及 domain，这样的结构使得查找更加迅速。